feat: create update and delete and list to entity challenge

+++ euduvido-api/src/main/java/com/euduvido/euduvido_api/application/usecases/DeleteUserUseCase.java
@@ -1,6 +1,9 @@
 package com.euduvido.euduvido_api.application.usecases;
 
+import com.euduvido.euduvido_api.domain.entities.User;
 import com.euduvido.euduvido_api.domain.repositories.UserRepository;
+
+import java.util.Optional;
 
 /**
  * Caso de uso: Deletar um usuário.
@@ -16,9 +19,17 @@ public class DeleteUserUseCase {
     /**
      * Executa a deleção de um usuário
      * @param id Nome do usuário
+     * @return void
      * @throws IllegalArgumentException se email já existe ou dados são inválidos
      */
     public void execute(Long id) {
+        // Verifica se o usuário existe
+        Optional<User> userOpt = userRepository.findById(id);
+
+        if (userOpt.isEmpty()) {
+            throw new IllegalArgumentException("Usuário não encontrado");
+        }
+
         userRepository.deleteById(id);
     }
 }
